fixing bool type allocations

+++ src/type_DeterminantString.jl
@@ -1,9 +1,9 @@
 using StaticArrays
 
 struct DeterminantString{N}
-    norbs::Int16
-    nelec::UInt8
+    norbs::Int
+    nelec::Int
     config::SVector{N, Int}
     #config::MVector{N, UInt}
-    index::UInt
+    index::Int
 end
